backend: refactoring y control de errores

+++ backend/routes/api.php
@@ -26,17 +26,19 @@ Route::post('login', [AuthController::class, 'login']);
 Route::get('users', [UserController::class, 'allusers']);
 Route::get('prueba', [ProductController::class, 'index']);
 
-// Ruta protegida que requiere autenticación
 Route::apiResource('products', ProductController::class);
 
-Route::put('/user/{id}', [UserController::class, 'update']);
-Route::delete('/user/{id}', [UserController::class, 'delete']);
 
 Route::middleware('auth:sanctum')->group(function () {
 
     Route::post('register', [AuthController::class, 'register'])->middleware('role:Administrador');
     Route::get('roles', [RoleController::class, 'index'])->middleware('role:Administrador');
-    Route::get('user', [AuthController::class, 'user']);
+    Route::put('user/{id}', [UserController::class, 'update'])->middleware('role:Administrador');
 
+    Route::delete('/user/{id}', [UserController::class, 'delete'])->middleware('role:Administrador');
+
+    Route::put('user/{id}/change-password', [UserController::class, 'changePassword'])->middleware('role:Administrador');
+
+    Route::get('user', [AuthController::class, 'user']);
     Route::post('logout', [AuthController::class, 'logout']);
 });
